backend/requirements.txt
alembic==1.13.1
fiona==1.9.5
Flask-CORS==6.0.1
Flask-SQLAlchemy==3.1.1
Flask==3.1.2
geopandas==0.14.1
matplotlib==3.10.7
nbformat>=4.2.0
numpy==1.26.4
pandas==2.3.2
plotly==6.3.1
pyproj==3.6.1
python-dotenv==0.19.2
scikit-learn==1.7.2
seaborn==0.13.2
shapely==2.0.2
SQLAlchemy==2.0.23
requests>=2.31.0
PyJWT>=2.8.0